Create CopasiModel class for encapsulation
It became apparent that the id management would introduce a lot of global variables. Also, creating functional reactions for subtraction needs intermediate species that needed to be returned by the function so that the main program can add it to the species list. This can be avoided using a class that encapsulates everything nicely.